Add .git-blame-ignore-revs
After migrating the code style to `black`, the output of `git blame`
will be uniformative as the author of a pure code refactoring is shown
instead of the original one. Adding the file `.git-blame-ignore-revs`
allows to ignore
changes made in listed commits when assigning blame. This requires a Git
version of at least 2.23. The command to run is
`git blame SOME_FILE.py --ignore-revs-file .git-blame-ignore-revs`.