Simplify inheritance of DynamicsRecovery
Running `super` with `DynamicsRecovery` and `self` specified causes
problems when code is changed and the class reinstantiated. The
TypeError `super(type, obj): obj must be an instance or subtype of type`
is thrown. When base
class and instantiated instance are not specified in `super`, the
class behavior is the unchanged but the problem fixed.