Add _base.py
The file is intended to host any code serving as base of further
implementations in `core/` of other submodules. For example, base
classes applicable to the majority of the code base are defined here. As
of now, the base class `DynamicsBase` is defined. It serves as the
foundation of implemented dynamical models.