Add core submodule (#373)
* Add submodule `core`

* Add submodule for `core/tests` for unit tests

* Add test_base.py

* Add composite `get_adata` strategy to genereate an AnnData object. The strategy is
unit tested through `TestAdataGeneration`.

* Add base class `TestBase`. The class includes the methods
`_subset_modalities` and `_convert_to_float`.

* Make `get_adata` and `TestBase` importable

* Add _anndata.py

The file is intended to host code related to manipulating AnnData
objects. As of now, the functions `get_modality`, `make_dense`,
`make_sparse` and `set_modality` are implemented.

* `get_modality` allows easily retrieving any modality from `X`,
`layers` as well as `obsm`.

* `make_dense` densifies a specified sparse AnnData entries.

* `make_sparse` makes specified AnnData entries sparse.

* `set_modality` allows setting a modality of an AnnData object to a new
value.

* Make functions from `_anndata.py` importable

* Add test_anndata.py

The file contains unit tests of functions defined in `_anndata.py`.

* Add _arithmetic.py

The file is intended to host code to perform arithemtic operations. As
of now, the functions `clipped_log`, `invert`, `prod_sum` and `sum` are
implemented.

* `clipped_log` logarithmizes an array with values clipped to a
specified interval.

* `invert` allows inverting an array with entries equal to zero.

* `prod_sum` implements muliplying two arrays and summing the result
along a given axis.

* `sum` sums an array (dense or sparse) over a given axis.

* Make functions from `_arithmetic.py` importable

* Add test_arithmetic.py

Add unit test for functions `clipped_log`, `invert`, `prod_sum` and
`sum` implemented in in `_arithmetic.py`.

* Add _linear_models.py

The file is inteded to host code to fit linear models. As of now, the
class `LinearRegression` is implemented. It largely follows the
implementation of `leastsq_NxN` in `scvelo/tools/optimization.py`.

* `LinearRegression` allows fitting a linear model per feature on
extreme quantiles.

* Make `LinearRegression` importable

* Add test_linear_models.py

Adds a unit test for `LinearRegression`.

* Add _metrics.py

The file is intended to host implementations of different metrics. As of
now, the l2 norm has been implemented.

* `l2_norm` takes the l2 norm of a given array (dense or sparse) along a
given axis.

* Make `l2_norm` importable

* Add test_metrics.py

Add unit test for implementation of l2 norm.

* Add _base.py

The file is intended to host any code serving as base of further
implementations in `core/` of other submodules. For example, base
classes applicable to the majority of the code base are defined here. As
of now, the base class `DynamicsBase` is defined. It serves as the
foundation of implemented dynamical models.

* Add _models.py

The file is intended to host implementations of dynamical systems. As of
now, the splicing model has been implemented.

* Make `SplicingDynamics` importable

* Add test_models.py

Add unit tests for SpicingDynamics.

* Update .gitignore

Ignore files generated by hypothesis.

* Update requirements-dev.txt

Add hypothesis.

* Update setup.py

Add hypothesis to dev installation.

* Update ci.yml

Install hypothesis prior to running unit tests.

* Update pytest.ini

* Scan only files of the form `test_*.py`.

* Add scvelo to testpaths.

* Deprecate leastsq_NxN

* Add deprecation notice to `leastsq_NxN`

* Replace the function `leastsq_NxN` in favor of the `LinearRegression`
class.

* Deprecate `sum_obs` and `sum_var`

Deprecate and replace functions in favor of `sum`.

* Deprecate `prod_sum_ods` and `prod_sum_var`

Functions are deprecated and replaced in favor of `prod_sum`.

* Deprecate `inv`

The function is deprecated and replaced in favor of `invert`.

* Deprecate `log`

Function is deprecated and replaced in favor of `clipped_log`.

* Deprecate `norm` and `vector_norm`

Functions are deprecated and replaced by `l2_norm`.

* Deprecate `mRNA`

The function is deprecated and replaced in favor of
`SplicingDynamics::get_solution`.